refactor: refactor paypal routes
